+++ RealmSwift-swift2.0/Realm.swift
@@ -470,8 +470,9 @@ public final class Realm {
     - parameter notificationToken: The token returned from `addNotificationBlock(_:)`
                                    corresponding to the notification block to remove.
     */
+    @available(*, deprecated=1, message="Use NotificationToken.stop()")
     public func removeNotification(notificationToken: NotificationToken) {
-        rlmRealm.removeNotification(notificationToken)
+        notificationToken.stop()
     }